// src/types/phase.ts
export enum Phase {
    Start = "Start",
    Lobby = "Lobby",
    Shuffle = "Shuffle",
    Deal = "Deal",
    PickOrPass = "PickOrPass",
    Discard = "Discard",
    ChooseSuit = "ChooseSuit",
    Play = "Play",
    CalculateWinner = "CalculateWinner",
    UpdateScore = "UpdateScore",
    End = "End",
}